+++ Hearthstone Deck Tracker/Windows/OverlayWindow.DeckLists.cs
@@ -26,9 +26,6 @@ namespace Hearthstone_Deck_Tracker.Windows
 					case DeckPanel.CardCounter:
 						StackPanelPlayer.Children.Add(CanvasPlayerCount);
 						break;
-					case DeckPanel.Fatigue:
-						StackPanelPlayer.Children.Add(LblPlayerFatigue);
-						break;
 					case DeckPanel.DeckTitle:
 						StackPanelPlayer.Children.Add(LblDeckTitle);
 						break;
@@ -63,9 +60,6 @@ namespace Hearthstone_Deck_Tracker.Windows
 						break;
 					case DeckPanel.CardCounter:
 						StackPanelOpponent.Children.Add(CanvasOpponentCount);
-						break;
-					case DeckPanel.Fatigue:
-						StackPanelOpponent.Children.Add(LblOpponentFatigue);
 						break;
 					case DeckPanel.Winrate:
 						StackPanelOpponent.Children.Add(LblWinRateAgainst);
@@ -109,27 +103,11 @@ namespace Hearthstone_Deck_Tracker.Windows
 			var fatigueDamage = Math.Max(_game.Opponent.Fatigue + 1, 1);
 			if(cardsLeftInDeck <= 0)
 			{
-				LblOpponentFatigue.Text = string.Format(
-					LocUtil.Get("Overlay_DeckList_Label_FatigueNextDraw"),
-					fatigueDamage
-				);
-
 				LblOpponentDrawChance2.Text = "0%";
 				LblOpponentDrawChance1.Text = "0%";
 				LblOpponentHandChance2.Text = cardCount <= 0 ? "0%" : "100%";
 				LblOpponentHandChance1.Text = cardCount <= 0 ? "0%" : "100%";
 				return;
-			}
-			else if(fatigueDamage > 1 || WotogCounterHelper.ShowOpponentFatigueCounter)
-			{
-				LblOpponentFatigue.Text = string.Format(
-					LocUtil.Get("Overlay_DeckList_Label_FatigueDamage"),
-					fatigueDamage
-				);
-			}
-			else
-			{
-				LblOpponentFatigue.Text = "";
 			}
 
 			var handWithoutCoin = cardCount - (_game.Opponent.HasCoin ? 1 : 0);
@@ -149,30 +127,6 @@ namespace Hearthstone_Deck_Tracker.Windows
 		{
 			LblCardCount.Text = cardCount.ToString();
 			LblDeckCount.Text = cardsLeftInDeck.ToString();
-
-			var fatigueDamage = Math.Max(_game.Player.Fatigue + 1, 1);
-			if(cardsLeftInDeck <= 0)
-			{
-				LblPlayerFatigue.Text = string.Format(
-					LocUtil.Get("Overlay_DeckList_Label_FatigueNextDraw"),
-					fatigueDamage
-				);
-
-				LblDrawChance2.Text = "0%";
-				LblDrawChance1.Text = "0%";
-				return;
-			}
-			else if(fatigueDamage > 1 || WotogCounterHelper.ShowPlayerFatigueCounter)
-			{
-				LblPlayerFatigue.Text = string.Format(
-					LocUtil.Get("Overlay_DeckList_Label_FatigueDamage"),
-					fatigueDamage
-				);
-			}
-			else
-			{
-				LblPlayerFatigue.Text = "";
-			}
 
 			var drawNextTurn2 = Math.Round(200.0f / cardsLeftInDeck, 1);
 			LblDrawChance2.Text = (cardsLeftInDeck == 1 ? 100 : drawNextTurn2) + "%";